ship: Remove topo_dirty flag.
This is part of preparing for zone_info. Bulk mutators of the ship are
expected to call rebuild_topology() when they are done; zone_info
tracking across such a rebuild might not be perfect.

Incremental mutators of the ship (surface add/remove) are expected to
call rebuild_topology() themselves if they give up on an incremental
update; and eventually to do their own zone_info tracking across such a
rebuild (when splitting a space, for example, we *know* the possible
subspaces afterward)

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>