main: fix double-free of rigid bodies
If the physics component's instance data was moved (to cover a deletion),
a stale copy of the data is left behind in the vacated slot. When another
component instance is allocated, it inherits that slot. Logic in mesher.cc
assumed that if the rb ptr was non-null, it already had an rb to update
(which is important for the chunk mesher itself). Later deleting this
component instance would double-free the rb object.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>