component: slim down physics_component data
We never actually used this, we just paid to move it around.
It didn't really make sense to store this per-entity anyway -- the physics
mesh itself isn't owned by the entity.

If we /do/ need entity-owned physics meshes, we'll add those later in a
new component (which will be quite a lot rarer)

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>