component: slim down entity->wire walking a bit
We can call attach_topo_find() on the attach itself; we don't need the
intermediate step of grabbing its data and the parent from that.

Exposes one more layer of wiring topo consolidation possibilities --
previously only the path from the parent to the wire representative
would be flattened; here we can go directly from the attach to the
wire.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>