main: Fix crash in play_state::rebuild_ui with entity deletion
If the use-key raycast succeeds, and this frame we delete that entity,
then we would have reached into freed memory in rebuild_ui when
determining the string to show the user.

Just rearrange things so that there's no entity deletion (currently done
by tool use) between the raycast and the rebuild_ui call.

While we're at it, make use_entity local to the play_state; there's no
good reason for this to be global

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>