main/tools: Support entities multiple blocks tall
This is still a /little/ hackish in that I think we'd prefer to
support multiple surface components for entities that are doing this,
or rework the entity/block-system code completely.

Allows block entities to be more than one block tall, as the door is.

Get rid of some spurious chunk rebuilds when doing entity-only work.

Move some more logic into destroy_entity that was previously spread
everywhere.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>