tools: Add a ghetto flashlight
The flashlight operates as a tool (which means if you change tools you will not
get updates). The implementation is a cheap trick which simply creates a light
where the flashlight's beam would have hit, and moves it as the player moves.

The main reason that it's ghetto is because its state is binary, either it hit
something and you have light, or it hit nothing and it seems as if it's not on.

At the flashlight's current throw, only the corners in the ceiling hit this
ghetto condition.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>