power_component: Replace power_override
Perhaps a better solution to avoid adding a new field to override the power tick
is to have a 0 powered component. Requiring and using 0 power. Something has to
change eventually to support a remote power source, and this seems as good of a
hack as power_override to me.

Keeping this patch separate since Caleb already saw and agreed to
power_override.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>