Update lightfield after tick_light_components
This change is primarily motivated to support the way the ghetto flashlight is
designed. It does have a potentially large impact on the various components
which emanate light, and I don't know the code well enough to determine whether
or not it actually matters to the other components.

The flashlight is toggled on and off via powered as opposed to intensity. I
believe this makes more sense in the long run because it allows alt_use to
control the intensity without requiring to save off the current intensity when
the light is turned off. Using powered, the only state needed to support lights
of different intensities based on alt_use() is requested_intensity, which is
already part of the "state". This causes a problem (I believe even if intensity
was used instead there is possible a similar problem).

The flashlight has some intelligence built into it whereby it tries to avoid
updating the lightfield when the lit block hasn't changed (and the flashlight
state hasn't changed). For example if you turn on the light and don't move,
nothing needs to happen (since there are no shadows and such). Before this
patch, the following can happen:

1. use flashlight (power.powered == true, mark block updates).
    At this point the light intensity is 0 because turning the flashlight
off turned off the power and thereby set the intensity to 0.
2. Sometime interval later, update_lightfield() is called as part of the update
loop.
    Intensity is still 0 here because we haven't yet done the tick on the
flashlight component.
3. Shortly after update_lightfield(), tick_light_components().
    Intensity is restored to the requested intensity.
4. flashlight_tool->preview() is called.
    player hasn't move and light state hasn't changed, so blocks are not
updated.

Full disclosure: I didn't go too far down the path of using intensity instead of
powered because I prefer using powered to control on/off. I am willing to
revisit this if there is some significant improvement to that strategy.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>