flashlight: Implement select
This serves as the opposite of the unselect() function. It is useful when state
needs to be preserved while switching tools without the player explicitly
"using" the tool.

The example with the flashlight is the player turns on the flashlight, needs to
use another tool so puts the flashlight in his/her pocket, but doesn't turn it
off. On pulling the flashlight out, the light is still on.

This behavior is arguably weird. I'll be cool with it if this patch is rejected.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>