flashlight: Pull out intensity (track locally)
This change makes intensity in the engine separate from the intensity (now named
brightness) of the flashlight. This will allow intensity controls to happen
independently of how the player uses the flashlight tool to control brightness.

Added here is the currently useless scale_brightness() so one can see how I
intend to do this.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>