flashlight: Clamp minimum brightness
A flashlight will always emit some amount of light unless you're in a blackhole.
Clamping the minimum brightness is a cheap way to achieve the effect. This makes
the flashlight /slightly/ less ghetto.

A likely better way to do this would be to detect when the flashlight's throw
doesn't hit a wall and move the light to the location of the player with a very
low intensity. That would probably simulate ambient light. Since it sounds like
the lighting is due for overhaul, I won't bother with this for now.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>